LPS-108943 Move listener to prevent links click to PageEditor
